Add output_type() to abstract outputs and handles
Expose output_type() through HasScriptPubkey so it is available on
abstract outputs. Add TxInHandle::output_type() to resolve input
types from prevouts when available.

Move shared output classification into primitives so these methods
live on the requested abstract and handle types without creating a
crate cycle.
