+++ src/crates/fingerprints/src/fingerprints/output.rs
@@ -1,17 +1,15 @@
-use tx_indexer_primitives::HasScriptPubkey;
-
-use crate::classify::classify_script_pubkey;
-use crate::types::OutputType;
+use tx_indexer_primitives::{HasScriptPubkey, OutputType};
 
 /// Classify an output by its scriptPubKey.
 pub fn output_type(output: &(impl HasScriptPubkey + ?Sized)) -> OutputType {
-    classify_script_pubkey(&output.script_pubkey_bytes())
+    output.output_type()
 }
 
 /// Bundled trait for output-level fingerprints.
 pub trait HasOutputFingerprints: HasScriptPubkey {
+    // TODO: add output level wallet fingerprints beyond scriptPubKey classification.
     fn output_type(&self) -> OutputType {
-        output_type(self)
+        HasScriptPubkey::output_type(self)
     }
 }
